+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Remember that the keywords that you can use are the following:
 
 ### enhancement
 - On-host: added apache and nginx agentTypes
+- On-host: added nri-postgresql agentType.
 - Added `agent-type validate --file <path>` subcommand to `newrelic-agent-control-cli` and `newrelic-agent-control-k8s-cli` for schema-level validation of agent type definition files.
 - Remove static regex validator for newrelic-infra Agent Type.